Separated declaration and implementation of integrationTest
